bharatSTR/requirements/core.txt
--extra-index-url https://download.pytorch.org/whl/cpu

aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.10.0
huggingface-hub==0.26.1
idna==3.10
jinja2==3.1.4
joblib==1.4.2
lightning-utilities==0.11.8
markupsafe==3.0.2
mpmath==1.3.0
multidict==6.1.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
packaging==24.1
pillow==11.0.0
propcache==0.2.0
pytorch-lightning==2.4.0
pyyaml==6.0.2
regex==2024.9.11
requests==2.32.3
safetensors==0.4.5
sympy==1.13.1
timm==1.0.11
torch==2.5.0+cpu
torchmetrics==1.5.1
torchvision==0.20.0+cpu
tqdm==4.66.5
typing-extensions==4.12.2
urllib3==2.2.3
yarl==1.16.0

# The following packages are considered to be unsafe in a requirements file:
setuptools==75.2.0
